+++ src/CoordinateOperation/Params/EPSG10508.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * PHPCoord.
+ *
+ * @author Doug Wright
+ */
+declare(strict_types=1);
+
+/** @internal */ return [
+  'geoidHeightCorrectionModelFile' => 'PHPCoord\\CoordinateOperation\\IGNFHeightRGF93v2NGFIGN78CorsicaProvider',
+];